ai_module/requirements.txt
streamlit==1.31.0
requests==2.31.0
openai==1.12.0
replicate==0.25.0
Pillow==10.2.0
google-generativeai==0.3.2
fal-client==0.4.1
google-cloud-aiplatform==1.38.0